Crate clang_sys

Rust bindings for libclang.

Documentation

Note that the documentation on https://docs.rs for this crate assumes usage of the runtime Cargo feature as well as the Cargo feature for the latest supported version of libclang (e.g., clang_11_0), neither of which are enabled by default.

Due to the usage of the runtime Cargo feature, this documentation will contain some additional types and functions to manage a dynamically loaded libclang instance at runtime.

Due to the usage of the Cargo feature for the latest supported version of libclang, this documentation will contain constants and functions that are not available in the oldest supported version of libclang (3.5). All of these types and functions have a documentation comment which specifies the minimum libclang version required to use the item.

Modules

clang_BlockCommandComment_getArgText

clang_BlockCommandComment_getCommandName

clang_BlockCommandComment_getNumArgs

clang_BlockCommandComment_getParagraph

clang_CXCursorSet_contains

clang_CXCursorSet_insert

clang_CXIndex_getGlobalOptions

clang_CXIndex_setGlobalOptions

clang_CXIndex_setInvocationEmissionPathOption

Only available on libclang 6.0 and later.

clang_CXXConstructor_isConvertingConstructor

Only available on libclang 3.9 and later.

clang_CXXConstructor_isCopyConstructor

Only available on libclang 3.9 and later.

clang_CXXConstructor_isDefaultConstructor

Only available on libclang 3.9 and later.

clang_CXXConstructor_isMoveConstructor

Only available on libclang 3.9 and later.

clang_CXXField_isMutable

Only available on libclang 3.8 and later.

clang_CXXMethod_isConst

clang_CXXMethod_isDefaulted

Only available on libclang 3.9 and later.

clang_CXXMethod_isPureVirtual

clang_CXXMethod_isStatic

clang_CXXMethod_isVirtual

clang_CXXRecord_isAbstract

Only available on libclang 6.0 and later.

clang_Comment_getChild

clang_Comment_getKind

clang_Comment_getNumChildren

clang_Comment_isWhitespace

clang_CompilationDatabase_dispose

clang_CompilationDatabase_fromDirectory

clang_CompilationDatabase_getAllCompileCommands

clang_CompilationDatabase_getCompileCommands

clang_CompileCommand_getArg

clang_CompileCommand_getDirectory

clang_CompileCommand_getFilename

Only available on libclang 3.8 and later.

clang_CompileCommand_getMappedSourceContent

Only available on libclang 3.8 and later.

clang_CompileCommand_getMappedSourcePath

Only available on libclang 3.8 and later.

clang_CompileCommand_getNumArgs

clang_CompileCommand_getNumMappedSources

clang_CompileCommands_dispose

clang_CompileCommands_getCommand

clang_CompileCommands_getSize

clang_Cursor_Evaluate

Only available on libclang 3.9 and later.

clang_Cursor_getArgument

clang_Cursor_getBriefCommentText

clang_Cursor_getCXXManglings

Only available on libclang 3.8 and later.

clang_Cursor_getCommentRange

clang_Cursor_getMangling

Only available on libclang 3.6 and later.

clang_Cursor_getModule

clang_Cursor_getNumArguments

clang_Cursor_getNumTemplateArguments

Only available on libclang 3.6 and later.

clang_Cursor_getObjCDeclQualifiers

clang_Cursor_getObjCManglings

Only available on libclang 6.0 and later.

clang_Cursor_getObjCPropertyAttributes

clang_Cursor_getObjCSelectorIndex

clang_Cursor_getOffsetOfField

Only available on libclang 3.7 and later.

clang_Cursor_getParsedComment

clang_Cursor_getRawCommentText

clang_Cursor_getReceiverType

clang_Cursor_getSpellingNameRange

clang_Cursor_getStorageClass

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentKind

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentType

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentUnsignedValue

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentValue

Only available on libclang 3.6 and later.

clang_Cursor_getTranslationUnit

clang_Cursor_hasAttrs

Only available on libclang 3.9 and later.

clang_Cursor_isAnonymous

Only available on libclang 3.7 and later.

clang_Cursor_isBitField

clang_Cursor_isDynamicCall

clang_Cursor_isExternalSymbol

Only available on libclang 5.0 and later.

clang_Cursor_isFunctionInlined

Only available on libclang 3.9 and later.

clang_Cursor_isMacroBuiltin

Only available on libclang 3.9 and later.

clang_Cursor_isMacroFunctionLike

Only available on libclang 3.9 and later.

clang_Cursor_isNull

clang_Cursor_isObjCOptional

clang_Cursor_isVariadic

clang_EnumDecl_isScoped

Only available on libclang 5.0 and later.

clang_EvalResult_dispose

Only available on libclang 3.9 and later.

clang_EvalResult_getAsDouble

Only available on libclang 3.9 and later.

clang_EvalResult_getAsInt

Only available on libclang 3.9 and later.

clang_EvalResult_getAsLongLong

Only available on libclang 4.0 and later.

clang_EvalResult_getAsStr

Only available on libclang 3.9 and later.

clang_EvalResult_getAsUnsigned

Only available on libclang 4.0 and later.

clang_EvalResult_getKind

Only available on libclang 3.9 and later.

clang_EvalResult_isUnsignedInt

Only available on libclang 4.0 and later.

clang_File_isEqual

Only available on libclang 3.6 and later.

clang_FullComment_getAsHTML

clang_FullComment_getAsXML

clang_HTMLStartTagComment_isSelfClosing

clang_HTMLStartTag_getAttrName

clang_HTMLStartTag_getAttrValue

clang_HTMLStartTag_getNumAttrs

clang_HTMLTagComment_getAsString

clang_HTMLTagComment_getTagName

clang_IndexAction_create

clang_IndexAction_dispose

clang_InlineCommandComment_getArgText

clang_InlineCommandComment_getCommandName

clang_InlineCommandComment_getNumArgs

clang_InlineCommandComment_getRenderKind

clang_InlineContentComment_hasTrailingNewline

clang_Location_isFromMainFile

clang_Location_isInSystemHeader

clang_Module_getASTFile

clang_Module_getFullName

clang_Module_getName

clang_Module_getNumTopLevelHeaders

clang_Module_getParent

clang_Module_getTopLevelHeader

clang_Module_isSystem

clang_ParamCommandComment_getDirection

clang_ParamCommandComment_getParamIndex

clang_ParamCommandComment_getParamName

clang_ParamCommandComment_isDirectionExplicit

clang_ParamCommandComment_isParamIndexValid

clang_Range_isNull

clang_TParamCommandComment_getDepth

clang_TParamCommandComment_getIndex

clang_TParamCommandComment_getParamName

clang_TParamCommandComment_isParamPositionValid

clang_TargetInfo_dispose

Only available on libclang 5.0 and later.

clang_TargetInfo_getPointerWidth

Only available on libclang 5.0 and later.

clang_TargetInfo_getTriple

Only available on libclang 5.0 and later.

clang_TextComment_getText

clang_Type_getAlignOf

clang_Type_getCXXRefQualifier

clang_Type_getClassType

clang_Type_getNamedType

Only available on libclang 3.9 and later.

clang_Type_getNumTemplateArguments

clang_Type_getObjCEncoding

Only available on libclang 3.9 and later.

clang_Type_getOffsetOf

clang_Type_getSizeOf

clang_Type_getTemplateArgumentAsType

clang_Type_isTransparentTagTypedef

Only available on libclang 5.0 and later.

clang_Type_visitFields

Only available on libclang 3.7 and later.

clang_VerbatimBlockLineComment_getText

clang_VerbatimLineComment_getText

clang_annotateTokens

clang_codeCompleteAt

clang_codeCompleteGetContainerKind

clang_codeCompleteGetContainerUSR

clang_codeCompleteGetContexts

clang_codeCompleteGetDiagnostic

clang_codeCompleteGetNumDiagnostics

clang_codeCompleteGetObjCSelector

clang_constructUSR_ObjCCategory

clang_constructUSR_ObjCClass

clang_constructUSR_ObjCIvar

clang_constructUSR_ObjCMethod

clang_constructUSR_ObjCProperty

clang_constructUSR_ObjCProtocol

clang_createCXCursorSet

clang_createIndex

clang_createTranslationUnit

clang_createTranslationUnit2

clang_createTranslationUnitFromSourceFile

clang_defaultCodeCompleteOptions

clang_defaultDiagnosticDisplayOptions

clang_defaultEditingTranslationUnitOptions

clang_defaultReparseOptions

clang_defaultSaveOptions

clang_disposeCXCursorSet

clang_disposeCXPlatformAvailability

clang_disposeCXTUResourceUsage

clang_disposeCodeCompleteResults

clang_disposeDiagnostic

clang_disposeDiagnosticSet

clang_disposeIndex

clang_disposeOverriddenCursors

clang_disposeSourceRangeList

clang_disposeString

clang_disposeStringSet

Only available on libclang 3.8 and later.

clang_disposeTokens

clang_disposeTranslationUnit

clang_enableStackTraces

clang_equalCursors

clang_equalLocations

clang_equalRanges

clang_equalTypes

clang_executeOnThread

clang_findIncludesInFile

clang_findReferencesInFile

clang_formatDiagnostic

clang_free

Only available on libclang 3.7 and later.

clang_getAddressSpace

Only available on libclang 5.0 and later.

clang_getAllSkippedRanges

Only available on libclang 4.0 and later.

clang_getArgType

clang_getArrayElementType

clang_getArraySize

clang_getCString

clang_getCXTUResourceUsage

clang_getCXXAccessSpecifier

clang_getCanonicalCursor

clang_getCanonicalType

clang_getChildDiagnostics

clang_getClangVersion

clang_getCompletionAnnotation

clang_getCompletionAvailability

clang_getCompletionBriefComment

clang_getCompletionChunkCompletionString

clang_getCompletionChunkKind

clang_getCompletionChunkText

clang_getCompletionNumAnnotations

clang_getCompletionParent

clang_getCompletionPriority

clang_getCursor

clang_getCursorAvailability

clang_getCursorCompletionString

clang_getCursorDefinition

clang_getCursorDisplayName

clang_getCursorExceptionSpecificationType

Only available on libclang 5.0 and later.

clang_getCursorExtent

clang_getCursorKind

clang_getCursorKindSpelling

clang_getCursorLanguage

clang_getCursorLexicalParent

clang_getCursorLinkage

clang_getCursorLocation

clang_getCursorPlatformAvailability

clang_getCursorReferenceNameRange

clang_getCursorReferenced

clang_getCursorResultType

clang_getCursorSemanticParent

clang_getCursorSpelling

clang_getCursorTLSKind

Only available on libclang 6.0 and later.

clang_getCursorType

clang_getCursorUSR

clang_getCursorVisibility

Only available on libclang 3.8 and later.

clang_getDeclObjCTypeEncoding

clang_getDefinitionSpellingAndExtent

clang_getDiagnostic

clang_getDiagnosticCategory

clang_getDiagnosticCategoryName

clang_getDiagnosticCategoryText

clang_getDiagnosticFixIt

clang_getDiagnosticInSet

clang_getDiagnosticLocation

clang_getDiagnosticNumFixIts

clang_getDiagnosticNumRanges

clang_getDiagnosticOption

clang_getDiagnosticRange

clang_getDiagnosticSetFromTU

clang_getDiagnosticSeverity

clang_getDiagnosticSpelling

clang_getElementType

clang_getEnumConstantDeclUnsignedValue

clang_getEnumConstantDeclValue

clang_getEnumDeclIntegerType

clang_getExceptionSpecificationType

Only available on libclang 5.0 and later.

clang_getExpansionLocation

clang_getFieldDeclBitWidth

clang_getFile

clang_getFileContents

Only available on libclang 6.0 and later.

clang_getFileLocation

clang_getFileName

clang_getFileTime

clang_getFileUniqueID

clang_getFunctionTypeCallingConv

clang_getIBOutletCollectionType

clang_getIncludedFile

clang_getInclusions

clang_getInstantiationLocation

clang_getLocation

clang_getLocationForOffset

clang_getModuleForFile

clang_getNullCursor

clang_getNullLocation

clang_getNullRange

clang_getNumArgTypes

clang_getNumCompletionChunks

clang_getNumDiagnostics

clang_getNumDiagnosticsInSet

clang_getNumElements

clang_getNumOverloadedDecls

clang_getOverloadedDecl

clang_getOverriddenCursors

clang_getPointeeType

clang_getPresumedLocation

clang_getRange

clang_getRangeEnd

clang_getRangeStart

clang_getRemappings

clang_getRemappingsFromFileList

clang_getResultType

clang_getSkippedRanges

clang_getSpecializedCursorTemplate

clang_getSpellingLocation

clang_getTUResourceUsageName

clang_getTemplateCursorKind

clang_getToken

clang_getTokenExtent

clang_getTokenKind

clang_getTokenLocation

clang_getTokenSpelling

clang_getTranslationUnitCursor

clang_getTranslationUnitSpelling

clang_getTranslationUnitTargetInfo

Only available on libclang 5.0 and later.

clang_getTypeDeclaration

clang_getTypeKindSpelling

clang_getTypeSpelling

clang_getTypedefDeclUnderlyingType

clang_getTypedefName

Only available on libclang 5.0 and later.

clang_hashCursor

clang_indexLoc_getCXSourceLocation

clang_indexLoc_getFileLocation

clang_indexSourceFile

clang_indexSourceFileFullArgv

Only available on libclang 3.8 and later.

clang_indexTranslationUnit

clang_index_getCXXClassDeclInfo

clang_index_getClientContainer

clang_index_getClientEntity

clang_index_getIBOutletCollectionAttrInfo

clang_index_getObjCCategoryDeclInfo

clang_index_getObjCContainerDeclInfo

clang_index_getObjCInterfaceDeclInfo

clang_index_getObjCPropertyDeclInfo

clang_index_getObjCProtocolRefListInfo

clang_index_isEntityObjCContainerKind

clang_index_setClientContainer

clang_index_setClientEntity

clang_isAttribute

clang_isConstQualifiedType

clang_isCursorDefinition

clang_isDeclaration

clang_isExpression

clang_isFileMultipleIncludeGuarded

clang_isFunctionTypeVariadic

clang_isInvalid

clang_isPODType

clang_isPreprocessing

clang_isReference

clang_isRestrictQualifiedType

clang_isStatement

clang_isTranslationUnit

clang_isUnexposed

clang_isVirtualBase

clang_isVolatileQualifiedType

clang_loadDiagnostics

clang_parseTranslationUnit

clang_parseTranslationUnit2

clang_parseTranslationUnit2FullArgv

Only available on libclang 3.8 and later.

clang_remap_dispose

clang_remap_getFilenames

clang_remap_getNumFiles

clang_reparseTranslationUnit

clang_saveTranslationUnit

clang_sortCodeCompletionResults

clang_suspendTranslationUnit

Only available on libclang 5.0 and later.

clang_toggleCrashRecovery

clang_tokenize

clang_visitChildren

support

Provides helper functionality.

Structs

CXCodeCompleteResults

CXComment

CXCompletionResult

CXCursor

CXCursorAndRangeVisitor

CXFileUniqueID

CXIdxAttrInfo

CXIdxBaseClassInfo

CXIdxCXXClassDeclInfo

CXIdxContainerInfo

CXIdxDeclInfo

CXIdxEntityInfo

CXIdxEntityRefInfo

CXIdxIBOutletCollectionAttrInfo

CXIdxImportedASTFileInfo

CXIdxIncludedFileInfo

CXIdxLoc

CXIdxObjCCategoryDeclInfo

CXIdxObjCContainerDeclInfo

CXIdxObjCInterfaceDeclInfo

CXIdxObjCPropertyDeclInfo

CXIdxObjCProtocolRefInfo

CXIdxObjCProtocolRefListInfo

CXPlatformAvailability

CXSourceLocation

CXSourceRange

CXSourceRangeList

CXString

CXStringSet

CXTUResourceUsage

CXTUResourceUsageEntry

CXToken

CXType

CXUnsavedFile

CXVersion

Functions

The set of functions loaded dynamically.

IndexerCallbacks

SharedLibrary

A dynamically loaded instance of the libclang library.

Enums

Version

The (minimum) version of a libclang shared library.

Constants

CXAvailability_Available

CXAvailability_Deprecated

CXAvailability_NotAccessible

CXAvailability_NotAvailable

CXBinaryOperator_Add

CXBinaryOperator_AddAssign

CXBinaryOperator_And

CXBinaryOperator_AndAssign

CXBinaryOperator_Assign

CXBinaryOperator_Cmp

CXBinaryOperator_Comma

CXBinaryOperator_Div

CXBinaryOperator_DivAssign

CXBinaryOperator_EQ

CXBinaryOperator_GE

CXBinaryOperator_GT

CXBinaryOperator_Invalid

CXBinaryOperator_LAnd

CXBinaryOperator_LE

CXBinaryOperator_LOr

CXBinaryOperator_LT

CXBinaryOperator_Mul

CXBinaryOperator_MulAssign

CXBinaryOperator_NE

CXBinaryOperator_Or

CXBinaryOperator_OrAssign

CXBinaryOperator_PtrMemD

CXBinaryOperator_PtrMemI

CXBinaryOperator_Rem

CXBinaryOperator_RemAssign

CXBinaryOperator_Shl

CXBinaryOperator_ShlAssign

CXBinaryOperator_Shr

CXBinaryOperator_ShrAssign

CXBinaryOperator_Sub

CXBinaryOperator_SubAssign

CXBinaryOperator_Xor

CXBinaryOperator_XorAssign

CXCallingConv_AAPCS

CXCallingConv_AAPCS_VFP

CXCallingConv_AArch64SVEPCS

Only produced by libclang 15.0 and later.

CXCallingConv_AArch64VectorCall

Only produced by libclang 8.0 and later.

CXCallingConv_C

CXCallingConv_Default

CXCallingConv_IntelOclBicc

CXCallingConv_Invalid

CXCallingConv_M68kRTD

Only produced by libclang 18.0 and later.

CXCallingConv_PreserveAll

Only produced by libclang 3.9 and later.

CXCallingConv_PreserveMost

Only produced by libclang 3.9 and later.

CXCallingConv_Swift

Only produced by libclang 3.9 and later.

CXCallingConv_SwiftAsync

Only produced by libclang 13.0 and later.

CXCallingConv_Unexposed

CXCallingConv_Win64

CXCallingConv_X86FastCall

CXCallingConv_X86Pascal

CXCallingConv_X86RegCall

Only produced by libclang 4.0 and later.

CXCallingConv_X86StdCall

CXCallingConv_X86ThisCall

CXCallingConv_X86VectorCall

Only produced by libclang 3.6 and later.

CXCallingConv_X86_64SysV

CXCallingConv_X86_64Win64

CXChildVisit_Break

CXChildVisit_Continue

CXChildVisit_Recurse

CXChoice_Default

CXChoice_Disabled

CXChoice_Enabled

CXCodeComplete_IncludeBriefComments

CXCodeComplete_IncludeCodePatterns

CXCodeComplete_IncludeCompletionsWithFixIts

CXCodeComplete_IncludeMacros

CXCodeComplete_SkipPreamble

CXCommentInlineCommandRenderKind_Bold

CXCommentInlineCommandRenderKind_Emphasized

CXCommentInlineCommandRenderKind_Monospaced

CXCommentInlineCommandRenderKind_Normal

CXCommentParamPassDirection_In

CXCommentParamPassDirection_InOut

CXCommentParamPassDirection_Out

CXComment_BlockCommand

CXComment_FullComment

CXComment_HTMLEndTag

CXComment_HTMLStartTag

CXComment_InlineCommand

CXComment_Null

CXComment_Paragraph

CXComment_ParamCommand

CXComment_TParamCommand

CXComment_Text

CXComment_VerbatimBlockCommand

CXComment_VerbatimBlockLine

CXComment_VerbatimLine

CXCompilationDatabase_CanNotLoadDatabase

CXCompilationDatabase_NoError

CXCompletionChunk_Colon

CXCompletionChunk_Comma

CXCompletionChunk_CurrentParameter

CXCompletionChunk_Equal

CXCompletionChunk_HorizontalSpace

CXCompletionChunk_Informative

CXCompletionChunk_LeftAngle

CXCompletionChunk_LeftBrace

CXCompletionChunk_LeftBracket

CXCompletionChunk_LeftParen

CXCompletionChunk_Optional

CXCompletionChunk_Placeholder

CXCompletionChunk_ResultType

CXCompletionChunk_RightAngle

CXCompletionChunk_RightBrace

CXCompletionChunk_RightBracket

CXCompletionChunk_RightParen

CXCompletionChunk_SemiColon

CXCompletionChunk_Text

CXCompletionChunk_TypedText

CXCompletionChunk_VerticalSpace

CXCompletionContext_AnyType

CXCompletionContext_AnyValue

CXCompletionContext_ArrowMemberAccess

CXCompletionContext_CXXClassTypeValue

CXCompletionContext_ClassTag

CXCompletionContext_DotMemberAccess

CXCompletionContext_EnumTag

CXCompletionContext_IncludedFile

CXCompletionContext_MacroName

CXCompletionContext_Namespace

CXCompletionContext_NaturalLanguage

CXCompletionContext_NestedNameSpecifier

CXCompletionContext_ObjCCategory

CXCompletionContext_ObjCClassMessage

CXCompletionContext_ObjCInstanceMessage

CXCompletionContext_ObjCInterface

CXCompletionContext_ObjCObjectValue

CXCompletionContext_ObjCPropertyAccess

CXCompletionContext_ObjCProtocol

CXCompletionContext_ObjCSelectorName

CXCompletionContext_ObjCSelectorValue

CXCompletionContext_StructTag

CXCompletionContext_Unexposed

CXCompletionContext_UnionTag

CXCompletionContext_Unknown

CXCursor_AddrLabelExpr

CXCursor_AlignedAttr

Only produced by libclang 9.0 and later.

CXCursor_AnnotateAttr

CXCursor_ArraySubscriptExpr

CXCursor_AsmLabelAttr

CXCursor_AsmStmt

Duplicate of CXCursor_GccAsmStmt.

CXCursor_BinaryOperator

CXCursor_BlockExpr

CXCursor_BreakStmt

CXCursor_BuiltinBitCastExpr

Only produced by ‘libclang’ 9.0 and later.

CXCursor_CStyleCastExpr

CXCursor_CUDAConstantAttr

CXCursor_CUDADeviceAttr

CXCursor_CUDAGlobalAttr

CXCursor_CUDAHostAttr

CXCursor_CUDASharedAttr

Only produced by libclang 3.6 and later.

CXCursor_CXXAccessSpecifier

CXCursor_CXXAddrspaceCastExpr

Only produced by libclang 12.0 and later.

CXCursor_CXXBaseSpecifier

CXCursor_CXXBoolLiteralExpr

CXCursor_CXXCatchStmt

CXCursor_CXXConstCastExpr

CXCursor_CXXDeleteExpr

CXCursor_CXXDynamicCastExpr

CXCursor_CXXFinalAttr

CXCursor_CXXForRangeStmt

CXCursor_CXXFunctionalCastExpr

CXCursor_CXXMethod

CXCursor_CXXNewExpr

CXCursor_CXXNullPtrLiteralExpr

CXCursor_CXXOverrideAttr

CXCursor_CXXParenListInitExpr

Only produced by libclang 16.0 and later.

CXCursor_CXXReinterpretCastExpr

CXCursor_CXXStaticCastExpr

CXCursor_CXXThisExpr

CXCursor_CXXThrowExpr

CXCursor_CXXTryStmt

CXCursor_CXXTypeidExpr

CXCursor_CallExpr

CXCursor_CaseStmt

CXCursor_CharacterLiteral

CXCursor_ClassDecl

CXCursor_ClassTemplate

CXCursor_ClassTemplatePartialSpecialization

CXCursor_CompoundAssignOperator

CXCursor_CompoundLiteralExpr

CXCursor_CompoundStmt

CXCursor_ConceptDecl

Only produced by libclang 15.0 and later.

CXCursor_ConceptSpecializationExpr

Only produced by libclang 15.0 and later.

CXCursor_ConditionalOperator

CXCursor_ConstAttr

CXCursor_Constructor

CXCursor_ContinueStmt

CXCursor_ConvergentAttr

Only produced by libclang 9.0 and later.

CXCursor_ConversionFunction

CXCursor_DLLExport

Only produced by libclang 3.8 and later.

CXCursor_DLLImport

Only produced by libclang 3.8 and later.

CXCursor_DeclRefExpr

CXCursor_DeclStmt

CXCursor_DefaultStmt

CXCursor_Destructor

CXCursor_DoStmt

CXCursor_EnumConstantDecl

CXCursor_EnumDecl

CXCursor_ExceptionSpecificationKind_BasicNoexcept

CXCursor_ExceptionSpecificationKind_ComputedNoexcept

CXCursor_ExceptionSpecificationKind_Dynamic

CXCursor_ExceptionSpecificationKind_DynamicNone

CXCursor_ExceptionSpecificationKind_MSAny

CXCursor_ExceptionSpecificationKind_None

CXCursor_ExceptionSpecificationKind_Unevaluated

CXCursor_ExceptionSpecificationKind_Uninstantiated

CXCursor_ExceptionSpecificationKind_Unparsed

CXCursor_FieldDecl

CXCursor_FixedPointLiteral

Only produced by libclang 7.0 and later.

CXCursor_FlagEnum

Only produced by libclang 8.0 and later.

CXCursor_FloatingLiteral

CXCursor_ForStmt

CXCursor_FriendDecl

Only produced by libclang 4.0 and later.

CXCursor_FunctionDecl

CXCursor_FunctionTemplate

CXCursor_GNUNullExpr

CXCursor_GenericSelectionExpr

CXCursor_GotoStmt

CXCursor_IBActionAttr

CXCursor_IBOutletAttr

CXCursor_IBOutletCollectionAttr

CXCursor_IfStmt

CXCursor_ImaginaryLiteral

CXCursor_InclusionDirective

CXCursor_IndirectGotoStmt

CXCursor_InitListExpr

CXCursor_IntegerLiteral

CXCursor_InvalidCode

CXCursor_InvalidFile

CXCursor_LabelRef

CXCursor_LabelStmt

CXCursor_LambdaExpr

CXCursor_LinkageSpec

CXCursor_MSAsmStmt

CXCursor_MacroDefinition

CXCursor_MacroExpansion

Duplicate of CXCursor_MacroInstantiation.

CXCursor_MemberRef

CXCursor_MemberRefExpr

CXCursor_ModuleImportDecl

CXCursor_NSConsumed

Only produced by libclang 8.0 and later.

CXCursor_NSConsumesSelf

Only produced by libclang 8.0 and later.

CXCursor_NSReturnsAutoreleased

Only produced by libclang 8.0 and later.

CXCursor_NSReturnsNotRetained

Only produced by libclang 8.0 and later.

CXCursor_NSReturnsRetained

Only produced by libclang 8.0 and later.

CXCursor_Namespace

CXCursor_NamespaceAlias

CXCursor_NamespaceRef

CXCursor_NoDeclFound

CXCursor_NoDuplicateAttr

CXCursor_NonTypeTemplateParameter

CXCursor_NotImplemented

CXCursor_NullStmt

CXCursor_OMPArraySectionExpr

Only produced by libclang 3.8 and later.

CXCursor_OMPArrayShapingExpr

Only produced by libclang 12.0 and later.

CXCursor_OMPAtomicDirective

Only produced by libclang 3.6 and later.

CXCursor_OMPBarrierDirective

CXCursor_OMPCancelDirective

Only produced by libclang 3.7 and later.

CXCursor_OMPCancellationPointDirective

Only produced by libclang 3.7 and later.

CXCursor_OMPCanonicalLoop

Only produced by libclang 13.0 and later.

CXCursor_OMPCriticalDirective

CXCursor_OMPDepobjDirective

Only produced by libclang 11.0 and later.

CXCursor_OMPDispatchDirective

Only produced by libclang 13.0 and later.

CXCursor_OMPDistributeDirective

Only produced by libclang 3.8 and later.

CXCursor_OMPDistributeParallelForDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPDistributeParallelForSimdDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPDistributeSimdDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPErrorDirective

Only produced by libclang 16.0 and later.

CXCursor_OMPFlushDirective

CXCursor_OMPForDirective

CXCursor_OMPForSimdDirective

Only produced by libclang 3.6 and later.

CXCursor_OMPGenericLoopDirective

Only produced by libclang 14.0 and later.

CXCursor_OMPInteropDirective

Only produced by libclang 13.0 and later.

CXCursor_OMPIteratorExpr

Only produced by libclang 12.0 and later.

CXCursor_OMPMaskedDirective

Only produced by libclang 13.0 and later.

CXCursor_OMPMaskedTaskLoopDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPMaskedTaskLoopSimdDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPMasterDirective

CXCursor_OMPMasterTaskLoopDirective

Only produced by libclang 10.0 and later.

CXCursor_OMPMasterTaskLoopSimdDirective

Only produced by libclang 10.0 and later.

CXCursor_OMPMetaDirective

Only produced by libclang 14.0 and later.

CXCursor_OMPOrderedDirective

Only produced by libclang 3.6 and later.

CXCursor_OMPParallelDirective

CXCursor_OMPParallelForDirective

CXCursor_OMPParallelForSimdDirective

Only produced by libclang 3.6 and later.

CXCursor_OMPParallelGenericLoopDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPParallelMaskedDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPParallelMaskedTaskLoopDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPParallelMaskedTaskLoopSimdDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPParallelMasterDirective

Only produced by libclang 10.0 and later.

CXCursor_OMPParallelMasterTaskLoopDirective

Only produced by libclang 10.0 and later.

CXCursor_OMPParallelMasterTaskLoopSimdDirective

Only produced by libclang 10.0 and later.

CXCursor_OMPParallelSectionsDirective

CXCursor_OMPScanDirective

Only produced by libclang 11.0 and later.

CXCursor_OMPScopeDirective

Only produced by libclang 18.0 and later.

CXCursor_OMPSectionDirective

CXCursor_OMPSectionsDirective

CXCursor_OMPSimdDirective

CXCursor_OMPSingleDirective

CXCursor_OMPTargetDataDirective

Only produced by libclang 3.8 and later.

CXCursor_OMPTargetDirective

Only produced by libclang 3.6 and later.

CXCursor_OMPTargetEnterDataDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPTargetExitDataDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPTargetParallelDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPTargetParallelForDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPTargetParallelForSimdDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPTargetParallelGenericLoopDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPTargetSimdDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTargetTeamsDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTargetTeamsDistributeDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTargetTeamsDistributeParallelForDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTargetTeamsDistributeParallelForSimdDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTargetTeamsDistributeSimdDirective

Only producer by libclang 4.0 and later.

CXCursor_OMPTargetTeamsGenericLoopDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPTargetUpdateDirective

Only produced by libclang 3.9 and later.

CXCursor_OMPTaskDirective

CXCursor_OMPTaskLoopDirective

Only produced by libclang 3.8 and later.

CXCursor_OMPTaskLoopSimdDirective

Only produced by libclang 3.8 and later.

CXCursor_OMPTaskgroupDirective

Only produced by libclang 3.7 and later.

CXCursor_OMPTaskwaitDirective

CXCursor_OMPTaskyieldDirective

CXCursor_OMPTeamsDirective

Only produced by libclang 3.6 and later.

CXCursor_OMPTeamsDistributeDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTeamsDistributeParallelForDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTeamsDistributeParallelForSimdDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTeamsDistributeSimdDirective

Only produced by libclang 4.0 and later.

CXCursor_OMPTeamsGenericLoopDirective

Only produced by libclang 15.0 and later.

CXCursor_OMPTileDirective

Only produced by libclang 13.0 and later.

CXCursor_OMPUnrollDirective

Only produced by libclang 13.0 and later.

CXCursor_ObjCAtCatchStmt

CXCursor_ObjCAtFinallyStmt

CXCursor_ObjCAtSynchronizedStmt

CXCursor_ObjCAtThrowStmt

CXCursor_ObjCAtTryStmt

CXCursor_ObjCAutoreleasePoolStmt

CXCursor_ObjCAvailabilityCheckExpr

Only produced by libclang 3.9 and later.

CXCursor_ObjCBoolLiteralExpr

CXCursor_ObjCBoxable

Only produced by libclang 8.0 and later.

CXCursor_ObjCBridgedCastExpr

CXCursor_ObjCCategoryDecl

CXCursor_ObjCCategoryImplDecl

CXCursor_ObjCClassMethodDecl

CXCursor_ObjCClassRef

CXCursor_ObjCDesignatedInitializer

Only produced by libclang 8.0 and later.

CXCursor_ObjCDynamicDecl

CXCursor_ObjCEncodeExpr

CXCursor_ObjCException

Only produced by libclang 8.0 and later.

CXCursor_ObjCExplicitProtocolImpl

Only produced by libclang 8.0 and later.

CXCursor_ObjCForCollectionStmt

CXCursor_ObjCImplementationDecl

CXCursor_ObjCIndependentClass

Only produced by libclang 8.0 and later.

CXCursor_ObjCInstanceMethodDecl

CXCursor_ObjCInterfaceDecl

CXCursor_ObjCIvarDecl

CXCursor_ObjCMessageExpr

CXCursor_ObjCNSObject

Only produced by libclang 8.0 and later.

CXCursor_ObjCPreciseLifetime

Only produced by libclang 8.0 and later.

CXCursor_ObjCPropertyDecl

CXCursor_ObjCProtocolDecl

CXCursor_ObjCProtocolExpr

CXCursor_ObjCProtocolRef

CXCursor_ObjCRequiresSuper

Only produced by libclang 8.0 and later.

CXCursor_ObjCReturnsInnerPointer

Only produced by libclang 8.0 and later.

CXCursor_ObjCRootClass

Only produced by libclang 8.0 and later.

CXCursor_ObjCRuntimeVisible

Only produced by libclang 8.0 and later.

CXCursor_ObjCSelectorExpr

CXCursor_ObjCSelfExpr

CXCursor_ObjCStringLiteral

CXCursor_ObjCSubclassingRestricted

Only produced by libclang 8.0 and later.

CXCursor_ObjCSuperClassRef

CXCursor_ObjCSynthesizeDecl

CXCursor_OverloadCandidate

Only produced by libclang 3.7 and later.

CXCursor_OverloadedDeclRef

CXCursor_PackExpansionExpr

CXCursor_PackedAttr

CXCursor_ParenExpr

CXCursor_ParmDecl

CXCursor_PreprocessingDirective

CXCursor_PureAttr

CXCursor_RequiresExpr

Only produced by libclang 15.0 and later.

CXCursor_ReturnStmt

CXCursor_SEHExceptStmt

CXCursor_SEHFinallyStmt

CXCursor_SEHLeaveStmt

CXCursor_SEHTryStmt

CXCursor_SizeOfPackExpr

CXCursor_StaticAssert

Only produced by libclang 3.9 and later.

CXCursor_StmtExpr

CXCursor_StringLiteral

CXCursor_StructDecl

CXCursor_SwitchStmt

CXCursor_TemplateRef

CXCursor_TemplateTemplateParameter

CXCursor_TemplateTypeParameter

CXCursor_TranslationUnit

CXCursor_TypeAliasDecl

CXCursor_TypeAliasTemplateDecl

Only produced by libclang 3.8 and later.

CXCursor_TypeRef

CXCursor_TypedefDecl

CXCursor_UnaryExpr

CXCursor_UnaryOperator

CXCursor_UnexposedAttr

CXCursor_UnexposedDecl

CXCursor_UnexposedExpr

CXCursor_UnexposedStmt

CXCursor_UnionDecl

CXCursor_UsingDeclaration

CXCursor_UsingDirective

CXCursor_VarDecl

CXCursor_VariableRef

CXCursor_VisibilityAttr

Only produced by libclang 3.8 and later.

CXCursor_WarnUnusedAttr

Only produced by libclang 9.0 and later.

CXCursor_WarnUnusedResultAttr

Only produced by libclang 9.0 and later.

CXCursor_WhileStmt

CXDiagnostic_DisplayCategoryId

CXDiagnostic_DisplayCategoryName

CXDiagnostic_DisplayColumn

CXDiagnostic_DisplayOption

CXDiagnostic_DisplaySourceLocation

CXDiagnostic_DisplaySourceRanges

CXDiagnostic_Error

CXDiagnostic_Fatal

CXDiagnostic_Ignored

CXDiagnostic_Note

CXDiagnostic_Warning

CXError_ASTReadError

CXError_Crashed

CXError_Failure

CXError_InvalidArguments

CXError_Success

CXEval_CFStr

CXEval_Float

CXEval_Int

CXEval_ObjCStrLiteral

CXEval_Other

CXEval_StrLiteral

CXEval_UnExposed

CXGlobalOpt_None

CXGlobalOpt_ThreadBackgroundPriorityForAll

CXGlobalOpt_ThreadBackgroundPriorityForEditing

CXGlobalOpt_ThreadBackgroundPriorityForIndexing

CXIdxAttr_IBAction

CXIdxAttr_IBOutlet

CXIdxAttr_IBOutletCollection

CXIdxAttr_Unexposed

CXIdxDeclFlag_Skipped

CXIdxEntityLang_C

CXIdxEntityLang_CXX

CXIdxEntityLang_None

CXIdxEntityLang_ObjC

CXIdxEntityLang_Swift

Only produced by libclang 5.0 and later.

CXIdxEntityRef_Direct

CXIdxEntityRef_Implicit

CXIdxEntity_CXXClass

CXIdxEntity_CXXConcept

Only produced by libclang 15.0 and later.

CXIdxEntity_CXXConstructor

CXIdxEntity_CXXConversionFunction

CXIdxEntity_CXXDestructor

CXIdxEntity_CXXInstanceMethod

CXIdxEntity_CXXInterface

CXIdxEntity_CXXNamespace

CXIdxEntity_CXXNamespaceAlias

CXIdxEntity_CXXStaticMethod

CXIdxEntity_CXXStaticVariable

CXIdxEntity_CXXTypeAlias

CXIdxEntity_Enum

CXIdxEntity_EnumConstant

CXIdxEntity_Field

CXIdxEntity_Function

CXIdxEntity_NonTemplate

CXIdxEntity_ObjCCategory

CXIdxEntity_ObjCClass

CXIdxEntity_ObjCClassMethod

CXIdxEntity_ObjCInstanceMethod

CXIdxEntity_ObjCIvar

CXIdxEntity_ObjCProperty

CXIdxEntity_ObjCProtocol

CXIdxEntity_Struct

CXIdxEntity_Template

CXIdxEntity_TemplatePartialSpecialization

CXIdxEntity_TemplateSpecialization

CXIdxEntity_Typedef

CXIdxEntity_Unexposed

CXIdxEntity_Union

CXIdxEntity_Variable

CXIdxObjCContainer_ForwardRef

CXIdxObjCContainer_Implementation

CXIdxObjCContainer_Interface

CXIndexOptIndexFunctionLocalSymbols

CXIndexOptIndexImplicitTemplateInstantiations

CXIndexOptNone

CXIndexOptSkipParsedBodiesInSession

CXIndexOptSuppressRedundantRefs

CXIndexOptSuppressWarnings

CXLanguage_C

CXLanguage_CPlusPlus

CXLanguage_Invalid

CXLanguage_ObjC

CXLinkage_External

CXLinkage_Internal

CXLinkage_Invalid

CXLinkage_NoLinkage

CXLinkage_UniqueExternal

CXLoadDiag_CannotLoad

CXLoadDiag_InvalidFile

CXLoadDiag_None

CXLoadDiag_Unknown

CXNameRange_WantQualifier

CXNameRange_WantSinglePiece

CXNameRange_WantTemplateArgs

CXObjCDeclQualifier_Bycopy

CXObjCDeclQualifier_Byref

CXObjCDeclQualifier_In

CXObjCDeclQualifier_Inout

CXObjCDeclQualifier_None

CXObjCDeclQualifier_Oneway

CXObjCDeclQualifier_Out

CXObjCPropertyAttr_assign

CXObjCPropertyAttr_atomic

CXObjCPropertyAttr_class

Only available on libclang 3.9 and later.

CXObjCPropertyAttr_copy

CXObjCPropertyAttr_getter

CXObjCPropertyAttr_noattr

CXObjCPropertyAttr_nonatomic

CXObjCPropertyAttr_readonly

CXObjCPropertyAttr_readwrite

CXObjCPropertyAttr_retain

CXObjCPropertyAttr_setter

CXObjCPropertyAttr_strong

CXObjCPropertyAttr_unsafe_unretained

CXObjCPropertyAttr_weak

CXPrintingPolicy_Alignof

CXPrintingPolicy_AnonymousTagLocations

CXPrintingPolicy_Bool

CXPrintingPolicy_ConstantArraySizeAsWritten

CXPrintingPolicy_ConstantsAsWritten

CXPrintingPolicy_FullyQualifiedName

CXPrintingPolicy_Half

CXPrintingPolicy_IncludeNewlines

CXPrintingPolicy_IncludeTagDefinition

CXPrintingPolicy_Indentation

CXPrintingPolicy_MSVCFormatting

CXPrintingPolicy_MSWChar

CXPrintingPolicy_PolishForDeclaration

CXPrintingPolicy_Restrict

CXPrintingPolicy_SuppressImplicitBase

CXPrintingPolicy_SuppressInitializers

CXPrintingPolicy_SuppressLifetimeQualifiers

CXPrintingPolicy_SuppressScope

CXPrintingPolicy_SuppressSpecifiers

CXPrintingPolicy_SuppressStrongLifetime

CXPrintingPolicy_SuppressTagKeyword

CXPrintingPolicy_SuppressTemplateArgsInCXXConstructors

CXPrintingPolicy_SuppressUnwrittenScope

CXPrintingPolicy_TerseOutput

CXPrintingPolicy_UnderscoreAlignof

CXPrintingPolicy_UseVoidForZeroParams

CXRefQualifier_LValue

CXRefQualifier_None

CXRefQualifier_RValue

CXReparse_None

CXResult_Invalid

CXResult_Success

CXResult_VisitBreak

CXSaveError_InvalidTU

CXSaveError_None

CXSaveError_TranslationErrors

CXSaveError_Unknown

CXSaveTranslationUnit_None

CXSymbolRole_AddressOf

CXSymbolRole_Call

CXSymbolRole_Declaration

CXSymbolRole_Definition

CXSymbolRole_Dynamic

CXSymbolRole_Implicit

CXSymbolRole_None

CXSymbolRole_Read

CXSymbolRole_Reference

CXSymbolRole_Write

CXTLS_Dynamic

CXTLS_None

CXTLS_Static

CXTUResourceUsage_AST

CXTUResourceUsage_AST_SideTables

CXTUResourceUsage_ExternalASTSource_Membuffer_MMap

CXTUResourceUsage_ExternalASTSource_Membuffer_Malloc

CXTUResourceUsage_GlobalCompletionResults

CXTUResourceUsage_Identifiers

CXTUResourceUsage_PreprocessingRecord

CXTUResourceUsage_Preprocessor

CXTUResourceUsage_Preprocessor_HeaderSearch

CXTUResourceUsage_Selectors

CXTUResourceUsage_SourceManagerContentCache

CXTUResourceUsage_SourceManager_DataStructures

CXTUResourceUsage_SourceManager_Membuffer_MMap

CXTUResourceUsage_SourceManager_Membuffer_Malloc

CXTemplateArgumentKind_Declaration

CXTemplateArgumentKind_Expression

CXTemplateArgumentKind_Integral

CXTemplateArgumentKind_Invalid

CXTemplateArgumentKind_Null

CXTemplateArgumentKind_NullPtr

CXTemplateArgumentKind_Pack

CXTemplateArgumentKind_Template

CXTemplateArgumentKind_TemplateExpansion

CXTemplateArgumentKind_Type

CXToken_Comment

CXToken_Identifier

CXToken_Keyword

CXToken_Literal

CXToken_Punctuation

CXTranslationUnit_CXXChainedPCH

CXTranslationUnit_CacheCompletionResults

CXTranslationUnit_CreatePreambleOnFirstParse

Only available on libclang 3.8 and later.

CXTranslationUnit_DetailedPreprocessingRecord

CXTranslationUnit_ForSerialization

CXTranslationUnit_IncludeBriefCommentsInCodeCompletion

CXTranslationUnit_Incomplete

CXTranslationUnit_KeepGoing

Only available on libclang 3.9 and later.

CXTranslationUnit_None

CXTranslationUnit_PrecompiledPreamble

CXTranslationUnit_SingleFileParse

Only available on libclang 5.0 and later.

CXTranslationUnit_SkipFunctionBodies

CXTypeLayoutError_Dependent

CXTypeLayoutError_Incomplete

CXTypeLayoutError_Invalid

CXTypeLayoutError_InvalidFieldName

CXTypeLayoutError_NotConstantSize

CXTypeLayoutError_Undeduced

Only produced by libclang 9.0 and later.

CXTypeNullability_Invalid

CXTypeNullability_NonNull

CXTypeNullability_Nullable

CXTypeNullability_NullableResult

Only produced by libclang 12.0 and later.

CXTypeNullability_Unspecified

CXType_Accum

Only produced by libclang 7.0 and later.

CXType_Atomic

Only produced by libclang 11.0 and later.

CXType_Attributed

Only produced by libclang 8.0 and later.

CXType_Auto

Only produced by libclang 3.8 and later.

CXType_BFloat16

Only produced by libclang 11.0 and later.

CXType_BTFTagAttributed

Only produced by libclang 15.0 and later.

CXType_BlockPointer

CXType_Bool

CXType_Char16

CXType_Char32

CXType_Char_S

CXType_Char_U

CXType_Complex

CXType_ConstantArray

CXType_Dependent

CXType_DependentSizedArray

CXType_Double

CXType_Elaborated

Only produced by libclang 3.9 and later.

CXType_Enum

CXType_ExtVector

Only produced by libclang 9.0 and later.

CXType_Float

CXType_Float16

Only produced by libclang 6.0 and later.

CXType_Float128

Only produced by libclang 3.9 and later.

CXType_FunctionNoProto

CXType_FunctionProto

CXType_Half

Only produced by libclang 5.0 and later.

CXType_Ibm128

Only produced by libclang 14.0 and later.

CXType_IncompleteArray

CXType_Int

CXType_Int128

CXType_Invalid

CXType_LValueReference

CXType_Long

CXType_LongAccum

Only produced by libclang 7.0 and later.

CXType_LongDouble

CXType_LongLong

CXType_MemberPointer

CXType_NullPtr

CXType_OCLEvent

Only produced by libclang 5.0 and later.

CXType_OCLImage1dArrayRO

Only produced by libclang 5.0 and later.

CXType_OCLImage1dArrayRW

Only produced by libclang 5.0 and later.

CXType_OCLImage1dArrayWO

Only produced by libclang 5.0 and later.

CXType_OCLImage1dBufferRO

Only produced by libclang 5.0 and later.

CXType_OCLImage1dBufferRW

Only produced by libclang 5.0 and later.

CXType_OCLImage1dBufferWO

Only produced by libclang 5.0 and later.

CXType_OCLImage1dRO

Only produced by libclang 5.0 and later.

CXType_OCLImage1dRW

Only produced by libclang 5.0 and later.

CXType_OCLImage1dWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayDepthRO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayDepthRW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayDepthWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayMSAADepthRO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayMSAADepthRW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayMSAADepthWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayMSAARO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayMSAARW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayMSAAWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayRO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayRW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dArrayWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dDepthRO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dDepthRW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dDepthWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dMSAADepthRO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dMSAADepthRW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dMSAADepthWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dMSAARO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dMSAARW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dMSAAWO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dRO

Only produced by libclang 5.0 and later.

CXType_OCLImage2dRW

Only produced by libclang 5.0 and later.

CXType_OCLImage2dWO

Only produced by libclang 5.0 and later.

CXType_OCLImage3dRO

Only produced by libclang 5.0 and later.

CXType_OCLImage3dRW

Only produced by libclang 5.0 and later.

CXType_OCLImage3dWO

Only produced by libclang 5.0 and later.

CXType_OCLIntelSubgroupAVCImeDualRefStreamin

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCImePayload

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCImeResult

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCImeResultDualRefStreamout

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCImeResultSingleRefStreamout

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCImeSingleRefStreamin

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCMcePayload

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCMceResult

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCRefPayload

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCRefResult

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCSicPayload

Only produced by libclang 8.0 and later.

CXType_OCLIntelSubgroupAVCSicResult

Only produced by libclang 8.0 and later.

CXType_OCLQueue

Only produced by libclang 5.0 and later.

CXType_OCLReserveID

Only produced by libclang 5.0 and later.

CXType_OCLSampler

Only produced by libclang 5.0 and later.

CXType_ObjCClass

CXType_ObjCId

CXType_ObjCInterface

CXType_ObjCObject

Only produced by libclang 8.0 and later.

CXType_ObjCObjectPointer

CXType_ObjCSel

CXType_ObjCTypeParam

Only produced by libclang 8.0 and later.

CXType_Overload

CXType_Pipe

Only produced by libclang 5.0 and later.

CXType_Pointer

CXType_RValueReference

CXType_Record

CXType_SChar

CXType_Short

CXType_ShortAccum

Only produced by libclang 7.0 and later.

CXType_Typedef

CXType_UAccum

Only produced by libclang 7.0 and later.

CXType_UChar

CXType_UInt

CXType_UInt128

CXType_ULong

CXType_ULongAccum

Only produced by libclang 7.0 and later.

CXType_ULongLong

CXType_UShort

CXType_UShortAccum

Only produced by libclang 7.0 and later.

CXType_Unexposed

CXType_VariableArray

CXType_Vector

CXType_Void

CXType_WChar

CXUnaryOperator_AddrOf

CXUnaryOperator_Coawait

CXUnaryOperator_Deref

CXUnaryOperator_Extension

CXUnaryOperator_Imag

CXUnaryOperator_Invalid

CXUnaryOperator_LNot

CXUnaryOperator_Minus

CXUnaryOperator_Not

CXUnaryOperator_Plus

CXUnaryOperator_PostDec

CXUnaryOperator_PostInc

CXUnaryOperator_PreDec

CXUnaryOperator_PreInc

CXUnaryOperator_Real

CXVisibility_Default

CXVisibility_Hidden

CXVisibility_Invalid

CXVisibility_Protected

CXVisit_Break

CXVisit_Continue

CX_CXXInvalidAccessSpecifier

CX_CXXPrivate

CX_CXXProtected

CX_CXXPublic

CX_SC_Auto

CX_SC_Extern

CX_SC_Invalid

CX_SC_None

CX_SC_OpenCLWorkGroupLocal

CX_SC_PrivateExtern

CX_SC_Register

CX_SC_Static

Functions

clang_BlockCommandComment_getArgText⚠

clang_BlockCommandComment_getCommandName⚠

clang_BlockCommandComment_getNumArgs⚠

clang_BlockCommandComment_getParagraph⚠

clang_CXCursorSet_contains⚠

clang_CXCursorSet_insert⚠

clang_CXIndex_getGlobalOptions⚠

clang_CXIndex_setGlobalOptions⚠

clang_CXIndex_setInvocationEmissionPathOption⚠

Only available on libclang 6.0 and later.

clang_CXXConstructor_isConvertingConstructor⚠

Only available on libclang 3.9 and later.

clang_CXXConstructor_isCopyConstructor⚠

Only available on libclang 3.9 and later.

clang_CXXConstructor_isDefaultConstructor⚠

Only available on libclang 3.9 and later.

clang_CXXConstructor_isMoveConstructor⚠

Only available on libclang 3.9 and later.

clang_CXXField_isMutable⚠

Only available on libclang 3.8 and later.

clang_CXXMethod_isConst⚠

clang_CXXMethod_isDefaulted⚠

Only available on libclang 3.9 and later.

clang_CXXMethod_isPureVirtual⚠

clang_CXXMethod_isStatic⚠

clang_CXXMethod_isVirtual⚠

clang_CXXRecord_isAbstract⚠

Only available on libclang 6.0 and later.

clang_Comment_getChild⚠

clang_Comment_getKind⚠

clang_Comment_getNumChildren⚠

clang_Comment_isWhitespace⚠

clang_CompilationDatabase_dispose⚠

clang_CompilationDatabase_fromDirectory⚠

clang_CompilationDatabase_getAllCompileCommands⚠

clang_CompilationDatabase_getCompileCommands⚠

clang_CompileCommand_getArg⚠

clang_CompileCommand_getDirectory⚠

clang_CompileCommand_getFilename⚠

Only available on libclang 3.8 and later.

clang_CompileCommand_getMappedSourceContent⚠

Only available on libclang 3.8 and later.

clang_CompileCommand_getMappedSourcePath⚠

Only available on libclang 3.8 and later.

clang_CompileCommand_getNumArgs⚠

clang_CompileCommand_getNumMappedSources⚠

clang_CompileCommands_dispose⚠

clang_CompileCommands_getCommand⚠

clang_CompileCommands_getSize⚠

clang_Cursor_Evaluate⚠

Only available on libclang 3.9 and later.

clang_Cursor_getArgument⚠

clang_Cursor_getBriefCommentText⚠

clang_Cursor_getCXXManglings⚠

Only available on libclang 3.8 and later.

clang_Cursor_getCommentRange⚠

clang_Cursor_getMangling⚠

Only available on libclang 3.6 and later.

clang_Cursor_getModule⚠

clang_Cursor_getNumArguments⚠

clang_Cursor_getNumTemplateArguments⚠

Only available on libclang 3.6 and later.

clang_Cursor_getObjCDeclQualifiers⚠

clang_Cursor_getObjCManglings⚠

Only available on libclang 6.0 and later.

clang_Cursor_getObjCPropertyAttributes⚠

clang_Cursor_getObjCSelectorIndex⚠

clang_Cursor_getOffsetOfField⚠

Only available on libclang 3.7 and later.

clang_Cursor_getParsedComment⚠

clang_Cursor_getRawCommentText⚠

clang_Cursor_getReceiverType⚠

clang_Cursor_getSpellingNameRange⚠

clang_Cursor_getStorageClass⚠

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentKind⚠

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentType⚠

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentUnsignedValue⚠

Only available on libclang 3.6 and later.

clang_Cursor_getTemplateArgumentValue⚠

Only available on libclang 3.6 and later.

clang_Cursor_getTranslationUnit⚠

clang_Cursor_hasAttrs⚠

Only available on libclang 3.9 and later.

clang_Cursor_isAnonymous⚠

Only available on libclang 3.7 and later.

clang_Cursor_isBitField⚠

clang_Cursor_isDynamicCall⚠

clang_Cursor_isExternalSymbol⚠

Only available on libclang 5.0 and later.

clang_Cursor_isFunctionInlined⚠

Only available on libclang 3.9 and later.

clang_Cursor_isMacroBuiltin⚠

Only available on libclang 3.9 and later.

clang_Cursor_isMacroFunctionLike⚠

Only available on libclang 3.9 and later.

clang_Cursor_isNull⚠

clang_Cursor_isObjCOptional⚠

clang_Cursor_isVariadic⚠

clang_EnumDecl_isScoped⚠

Only available on libclang 5.0 and later.

clang_EvalResult_dispose⚠

Only available on libclang 3.9 and later.

clang_EvalResult_getAsDouble⚠

Only available on libclang 3.9 and later.

clang_EvalResult_getAsInt⚠

Only available on libclang 3.9 and later.

clang_EvalResult_getAsLongLong⚠

Only available on libclang 4.0 and later.

clang_EvalResult_getAsStr⚠

Only available on libclang 3.9 and later.

clang_EvalResult_getAsUnsigned⚠

Only available on libclang 4.0 and later.

clang_EvalResult_getKind⚠

Only available on libclang 3.9 and later.

clang_EvalResult_isUnsignedInt⚠

Only available on libclang 4.0 and later.

clang_File_isEqual⚠

Only available on libclang 3.6 and later.

clang_FullComment_getAsHTML⚠

clang_FullComment_getAsXML⚠

clang_HTMLStartTagComment_isSelfClosing⚠

clang_HTMLStartTag_getAttrName⚠

clang_HTMLStartTag_getAttrValue⚠

clang_HTMLStartTag_getNumAttrs⚠

clang_HTMLTagComment_getAsString⚠

clang_HTMLTagComment_getTagName⚠

clang_IndexAction_create⚠

clang_IndexAction_dispose⚠

clang_InlineCommandComment_getArgText⚠

clang_InlineCommandComment_getCommandName⚠

clang_InlineCommandComment_getNumArgs⚠

clang_InlineCommandComment_getRenderKind⚠

clang_InlineContentComment_hasTrailingNewline⚠

clang_Location_isFromMainFile⚠

clang_Location_isInSystemHeader⚠

clang_Module_getASTFile⚠

clang_Module_getFullName⚠

clang_Module_getName⚠

clang_Module_getNumTopLevelHeaders⚠

clang_Module_getParent⚠

clang_Module_getTopLevelHeader⚠

clang_Module_isSystem⚠

clang_ParamCommandComment_getDirection⚠

clang_ParamCommandComment_getParamIndex⚠

clang_ParamCommandComment_getParamName⚠

clang_ParamCommandComment_isDirectionExplicit⚠

clang_ParamCommandComment_isParamIndexValid⚠

clang_Range_isNull⚠

clang_TParamCommandComment_getDepth⚠

clang_TParamCommandComment_getIndex⚠

clang_TParamCommandComment_getParamName⚠

clang_TParamCommandComment_isParamPositionValid⚠

clang_TargetInfo_dispose⚠

Only available on libclang 5.0 and later.

clang_TargetInfo_getPointerWidth⚠

Only available on libclang 5.0 and later.

clang_TargetInfo_getTriple⚠

Only available on libclang 5.0 and later.

clang_TextComment_getText⚠

clang_Type_getAlignOf⚠

clang_Type_getCXXRefQualifier⚠

clang_Type_getClassType⚠

clang_Type_getNamedType⚠

Only available on libclang 3.9 and later.

clang_Type_getNumTemplateArguments⚠

clang_Type_getObjCEncoding⚠

Only available on libclang 3.9 and later.

clang_Type_getOffsetOf⚠

clang_Type_getSizeOf⚠

clang_Type_getTemplateArgumentAsType⚠

clang_Type_isTransparentTagTypedef⚠

Only available on libclang 5.0 and later.

clang_Type_visitFields⚠

Only available on libclang 3.7 and later.

clang_VerbatimBlockLineComment_getText⚠

clang_VerbatimLineComment_getText⚠

clang_annotateTokens⚠

clang_codeCompleteAt⚠

clang_codeCompleteGetContainerKind⚠

clang_codeCompleteGetContainerUSR⚠

clang_codeCompleteGetContexts⚠

clang_codeCompleteGetDiagnostic⚠

clang_codeCompleteGetNumDiagnostics⚠

clang_codeCompleteGetObjCSelector⚠

clang_constructUSR_ObjCCategory⚠

clang_constructUSR_ObjCClass⚠

clang_constructUSR_ObjCIvar⚠

clang_constructUSR_ObjCMethod⚠

clang_constructUSR_ObjCProperty⚠

clang_constructUSR_ObjCProtocol⚠

clang_createCXCursorSet⚠

clang_createIndex⚠

clang_createTranslationUnit⚠

clang_createTranslationUnit2⚠

clang_createTranslationUnitFromSourceFile⚠

clang_defaultCodeCompleteOptions⚠

clang_defaultDiagnosticDisplayOptions⚠

clang_defaultEditingTranslationUnitOptions⚠

clang_defaultReparseOptions⚠

clang_defaultSaveOptions⚠

clang_disposeCXCursorSet⚠

clang_disposeCXPlatformAvailability⚠

clang_disposeCXTUResourceUsage⚠

clang_disposeCodeCompleteResults⚠

clang_disposeDiagnostic⚠

clang_disposeDiagnosticSet⚠

clang_disposeIndex⚠

clang_disposeOverriddenCursors⚠

clang_disposeSourceRangeList⚠

clang_disposeString⚠

clang_disposeStringSet⚠

Only available on libclang 3.8 and later.

clang_disposeTokens⚠

clang_disposeTranslationUnit⚠

clang_enableStackTraces⚠

clang_equalCursors⚠

clang_equalLocations⚠

clang_equalRanges⚠

clang_equalTypes⚠

clang_executeOnThread⚠

clang_findIncludesInFile⚠

clang_findReferencesInFile⚠

clang_formatDiagnostic⚠

clang_free⚠

Only available on libclang 3.7 and later.

clang_getAddressSpace⚠

Only available on libclang 5.0 and later.

clang_getAllSkippedRanges⚠

Only available on libclang 4.0 and later.

clang_getArgType⚠

clang_getArrayElementType⚠

clang_getArraySize⚠

clang_getCString⚠

clang_getCXTUResourceUsage⚠

clang_getCXXAccessSpecifier⚠

clang_getCanonicalCursor⚠

clang_getCanonicalType⚠

clang_getChildDiagnostics⚠

clang_getClangVersion⚠

clang_getCompletionAnnotation⚠

clang_getCompletionAvailability⚠

clang_getCompletionBriefComment⚠

clang_getCompletionChunkCompletionString⚠

clang_getCompletionChunkKind⚠

clang_getCompletionChunkText⚠

clang_getCompletionNumAnnotations⚠

clang_getCompletionParent⚠

clang_getCompletionPriority⚠

clang_getCursor⚠

clang_getCursorAvailability⚠

clang_getCursorCompletionString⚠

clang_getCursorDefinition⚠

clang_getCursorDisplayName⚠

clang_getCursorExceptionSpecificationType⚠

Only available on libclang 5.0 and later.

clang_getCursorExtent⚠

clang_getCursorKind⚠

clang_getCursorKindSpelling⚠

clang_getCursorLanguage⚠

clang_getCursorLexicalParent⚠

clang_getCursorLinkage⚠

clang_getCursorLocation⚠

clang_getCursorPlatformAvailability⚠

clang_getCursorReferenceNameRange⚠

clang_getCursorReferenced⚠

clang_getCursorResultType⚠

clang_getCursorSemanticParent⚠

clang_getCursorSpelling⚠

clang_getCursorTLSKind⚠

Only available on libclang 6.0 and later.

clang_getCursorType⚠

clang_getCursorUSR⚠

clang_getCursorVisibility⚠

Only available on libclang 3.8 and later.

clang_getDeclObjCTypeEncoding⚠

clang_getDefinitionSpellingAndExtent⚠

clang_getDiagnostic⚠

clang_getDiagnosticCategory⚠

clang_getDiagnosticCategoryName⚠

clang_getDiagnosticCategoryText⚠

clang_getDiagnosticFixIt⚠

clang_getDiagnosticInSet⚠

clang_getDiagnosticLocation⚠

clang_getDiagnosticNumFixIts⚠

clang_getDiagnosticNumRanges⚠

clang_getDiagnosticOption⚠

clang_getDiagnosticRange⚠

clang_getDiagnosticSetFromTU⚠

clang_getDiagnosticSeverity⚠

clang_getDiagnosticSpelling⚠

clang_getElementType⚠

clang_getEnumConstantDeclUnsignedValue⚠

clang_getEnumConstantDeclValue⚠

clang_getEnumDeclIntegerType⚠

clang_getExceptionSpecificationType⚠

Only available on libclang 5.0 and later.

clang_getExpansionLocation⚠

clang_getFieldDeclBitWidth⚠

clang_getFile⚠

clang_getFileContents⚠

Only available on libclang 6.0 and later.

clang_getFileLocation⚠

clang_getFileName⚠

clang_getFileTime⚠

clang_getFileUniqueID⚠

clang_getFunctionTypeCallingConv⚠

clang_getIBOutletCollectionType⚠

clang_getIncludedFile⚠

clang_getInclusions⚠

clang_getInstantiationLocation⚠

clang_getLocation⚠

clang_getLocationForOffset⚠

clang_getModuleForFile⚠

clang_getNullCursor⚠

clang_getNullLocation⚠

clang_getNullRange⚠

clang_getNumArgTypes⚠

clang_getNumCompletionChunks⚠

clang_getNumDiagnostics⚠

clang_getNumDiagnosticsInSet⚠

clang_getNumElements⚠

clang_getNumOverloadedDecls⚠

clang_getOverloadedDecl⚠

clang_getOverriddenCursors⚠

clang_getPointeeType⚠

clang_getPresumedLocation⚠

clang_getRange⚠

clang_getRangeEnd⚠

clang_getRangeStart⚠

clang_getRemappings⚠

clang_getRemappingsFromFileList⚠

clang_getResultType⚠

clang_getSkippedRanges⚠

clang_getSpecializedCursorTemplate⚠

clang_getSpellingLocation⚠

clang_getTUResourceUsageName⚠

clang_getTemplateCursorKind⚠

clang_getToken⚠

clang_getTokenExtent⚠

clang_getTokenKind⚠

clang_getTokenLocation⚠

clang_getTokenSpelling⚠

clang_getTranslationUnitCursor⚠

clang_getTranslationUnitSpelling⚠

clang_getTranslationUnitTargetInfo⚠

Only available on libclang 5.0 and later.

clang_getTypeDeclaration⚠

clang_getTypeKindSpelling⚠

clang_getTypeSpelling⚠

clang_getTypedefDeclUnderlyingType⚠

clang_getTypedefName⚠

Only available on libclang 5.0 and later.

clang_hashCursor⚠

clang_indexLoc_getCXSourceLocation⚠

clang_indexLoc_getFileLocation⚠

clang_indexSourceFile⚠

clang_indexSourceFileFullArgv⚠

Only available on libclang 3.8 and later.

clang_indexTranslationUnit⚠

clang_index_getCXXClassDeclInfo⚠

clang_index_getClientContainer⚠

clang_index_getClientEntity⚠

clang_index_getIBOutletCollectionAttrInfo⚠

clang_index_getObjCCategoryDeclInfo⚠

clang_index_getObjCContainerDeclInfo⚠

clang_index_getObjCInterfaceDeclInfo⚠

clang_index_getObjCPropertyDeclInfo⚠

clang_index_getObjCProtocolRefListInfo⚠

clang_index_isEntityObjCContainerKind⚠

clang_index_setClientContainer⚠

clang_index_setClientEntity⚠

clang_isAttribute⚠

clang_isConstQualifiedType⚠

clang_isCursorDefinition⚠

clang_isDeclaration⚠

clang_isExpression⚠

clang_isFileMultipleIncludeGuarded⚠

clang_isFunctionTypeVariadic⚠

clang_isInvalid⚠

clang_isPODType⚠

clang_isPreprocessing⚠

clang_isReference⚠

clang_isRestrictQualifiedType⚠

clang_isStatement⚠

clang_isTranslationUnit⚠

clang_isUnexposed⚠

clang_isVirtualBase⚠

clang_isVolatileQualifiedType⚠

clang_loadDiagnostics⚠

clang_parseTranslationUnit⚠

clang_parseTranslationUnit2⚠

clang_parseTranslationUnit2FullArgv⚠

Only available on libclang 3.8 and later.

clang_remap_dispose⚠

clang_remap_getFilenames⚠

clang_remap_getNumFiles⚠

clang_reparseTranslationUnit⚠

clang_saveTranslationUnit⚠

clang_sortCodeCompletionResults⚠

clang_suspendTranslationUnit⚠

Only available on libclang 5.0 and later.

clang_toggleCrashRecovery⚠

clang_tokenize⚠

clang_visitChildren⚠

get_library

Returns the library instance stored in TLS.

is_loaded

Returns whether a libclang shared library is loaded on this thread.

load

Loads a libclang shared library for use in the current thread.

load_manually

Loads a libclang shared library and returns the library instance.

set_library

Sets the library instance stored in TLS and returns the previous library.

unload

Unloads the libclang shared library in use in the current thread.

Type Aliases

CXAvailabilityKind

CXBinaryOperatorKind

CXCallingConv

CXChildVisitResult

CXChoice

CXClientData

CXCodeComplete_Flags

CXCommentInlineCommandRenderKind

CXCommentKind

CXCommentParamPassDirection

CXCompilationDatabase

CXCompilationDatabase_Error

CXCompileCommand

CXCompileCommands

CXCompletionChunkKind

CXCompletionContext

CXCompletionString

CXCursorKind

CXCursorSet

CXCursorVisitor

CXCursor_ExceptionSpecificationKind

CXDiagnostic

CXDiagnosticDisplayOptions

CXDiagnosticSet

CXDiagnosticSeverity

CXErrorCode

CXEvalResult

CXEvalResultKind

CXFieldVisitor

CXFile

CXGlobalOptFlags

CXIdxAttrKind

CXIdxClientASTFile

CXIdxClientContainer

CXIdxClientEntity

CXIdxClientFile

CXIdxDeclInfoFlags

CXIdxEntityCXXTemplateKind

CXIdxEntityKind

CXIdxEntityLanguage

CXIdxEntityRefKind

CXIdxObjCContainerKind

CXInclusionVisitor

CXIndex

CXIndexAction

CXIndexOptFlags

CXLanguageKind

CXLinkageKind

CXLoadDiag_Error

CXModule

CXNameRefFlags

CXObjCDeclQualifierKind

CXObjCPropertyAttrKind

CXPrintingPolicyProperty

CXRefQualifierKind

CXRemapping

CXReparse_Flags

CXResult

CXSaveError

CXSaveTranslationUnit_Flags

CXSymbolRole

CXTLSKind

CXTUResourceUsageKind

CXTargetInfo

CXTemplateArgumentKind

CXTokenKind

CXTranslationUnit

CXTranslationUnit_Flags

CXTypeKind

CXTypeLayoutError

CXTypeNullabilityKind

CXUnaryOperatorKind

CXVisibilityKind

CXVisitorResult

CX_CXXAccessSpecifier

CX_StorageClass