Struct clang_sys::support::Clang

pub struct Clang {
    pub path: PathBuf,
    pub version: Option<CXVersion>,
    pub c_search_paths: Vec<PathBuf>,
    pub cpp_search_paths: Vec<PathBuf>,
}

A clang executable.

Fields
path: PathBuf

The path to this clang executable.

version: Option<CXVersion>

The version of this clang executable if it could be parsed.

c_search_paths: Vec<PathBuf>

The directories searched by this clang executable for C headers.

cpp_search_paths: Vec<PathBuf>

The directories searched by this clang executable for C++ headers.

Methods

impl Clang

fn find(path: Option<&Path>) -> Option<Clang>

Returns a clang executable if one can be found.

If the CLANG_PATH environment variable is set, that is the instance of clang used. Otherwise, a series of directories are searched. First, If a path is supplied, that is the first directory searched. Then, the directory returned by llvm-config --bindir is searched. On OS X systems, xcodebuild -find clang will next be queried. Last, the directories in the system's PATH are searched.