Crate clang_sys

Rust bindings for libclang.

Supported Versions

3.5 - Documentation

3.6 - Documentation

3.7 - Documentation

3.8 - Documentation

3.9 - Documentation

Modules

support
Provides helper functionality.

Structs

CXCodeCompleteResults

CXCodeComplete_Flags

CXComment

CXCompilationDatabase

CXCompileCommand

CXCompileCommands

CXCompletionContext

CXCompletionResult

CXCompletionString

CXCursor

CXCursorAndRangeVisitor

CXCursorSet

CXDiagnostic

CXDiagnosticDisplayOptions

CXDiagnosticSet

CXFile

CXFileUniqueID

CXGlobalOptFlags

CXIdxAttrInfo

CXIdxBaseClassInfo

CXIdxCXXClassDeclInfo

CXIdxClientASTFile

CXIdxClientContainer

CXIdxClientEntity

CXIdxClientFile

CXIdxContainerInfo

CXIdxDeclInfo

CXIdxDeclInfoFlags

CXIdxEntityInfo

CXIdxEntityRefInfo

CXIdxIBOutletCollectionAttrInfo

CXIdxImportedASTFileInfo

CXIdxIncludedFileInfo

CXIdxLoc

CXIdxObjCCategoryDeclInfo

CXIdxObjCContainerDeclInfo

CXIdxObjCInterfaceDeclInfo

CXIdxObjCPropertyDeclInfo

CXIdxObjCProtocolRefInfo

CXIdxObjCProtocolRefListInfo

CXIndex

CXIndexAction

CXIndexOptFlags

CXModule

CXNameRefFlags

CXObjCDeclQualifierKind

CXObjCPropertyAttrKind

CXPlatformAvailability

CXRemapping

CXReparse_Flags

CXSaveTranslationUnit_Flags

CXSourceLocation

CXSourceRange

CXSourceRangeList

CXString

CXStringSet

CXTUResourceUsage

CXTUResourceUsageEntry

CXToken

CXTranslationUnit

CXTranslationUnit_Flags

CXType

CXUnsavedFile

CXVersion

IndexerCallbacks

Enums

CXAvailabilityKind

CXCallingConv

CXChildVisitResult

CXCommentInlineCommandRenderKind

CXCommentKind

CXCommentParamPassDirection

CXCompilationDatabase_Error

CXCompletionChunkKind

CXCursorKind

CXDiagnosticSeverity

CXErrorCode

CXIdxAttrKind

CXIdxEntityCXXTemplateKind

CXIdxEntityKind

CXIdxEntityLanguage

CXIdxEntityRefKind

CXIdxObjCContainerKind

CXLanguageKind

CXLinkageKind

CXLoadDiag_Error

CXRefQualifierKind

CXResult

CXSaveError

CXTUResourceUsageKind

CXTemplateArgumentKind

CXTokenKind

CXTypeKind

CXTypeLayoutError

CXVisibilityKind

CXVisitorResult

CX_CXXAccessSpecifier

CX_StorageClass

Constants

CXCodeComplete_IncludeBriefComments

CXCodeComplete_IncludeCodePatterns

CXCodeComplete_IncludeMacros

CXCompletionContext_AnyType

CXCompletionContext_AnyValue

CXCompletionContext_ArrowMemberAccess

CXCompletionContext_CXXClassTypeValue

CXCompletionContext_ClassTag

CXCompletionContext_DotMemberAccess

CXCompletionContext_EnumTag

CXCompletionContext_MacroName

CXCompletionContext_Namespace

CXCompletionContext_NaturalLanguage

CXCompletionContext_NestedNameSpecifier

CXCompletionContext_ObjCCategory

CXCompletionContext_ObjCClassMessage

CXCompletionContext_ObjCInstanceMessage

CXCompletionContext_ObjCInterface

CXCompletionContext_ObjCObjectValue

CXCompletionContext_ObjCPropertyAccess

CXCompletionContext_ObjCProtocol

CXCompletionContext_ObjCSelectorName

CXCompletionContext_ObjCSelectorValue

CXCompletionContext_StructTag

CXCompletionContext_Unexposed

CXCompletionContext_UnionTag

CXCompletionContext_Unknown

CXDiagnostic_DisplayCategoryId

CXDiagnostic_DisplayCategoryName

CXDiagnostic_DisplayColumn

CXDiagnostic_DisplayOption

CXDiagnostic_DisplaySourceLocation

CXDiagnostic_DisplaySourceRanges

CXGlobalOpt_None

CXGlobalOpt_ThreadBackgroundPriorityForAll

CXGlobalOpt_ThreadBackgroundPriorityForEditing

CXGlobalOpt_ThreadBackgroundPriorityForIndexing

CXIdxDeclFlag_Skipped

CXIndexOptIndexFunctionLocalSymbols

CXIndexOptIndexImplicitTemplateInstantiations

CXIndexOptNone

CXIndexOptSkipParsedBodiesInSession

CXIndexOptSuppressRedundantRefs

CXIndexOptSuppressWarnings

CXNameRange_WantQualifier

CXNameRange_WantSinglePiece

CXNameRange_WantTemplateArgs

CXObjCDeclQualifier_Bycopy

CXObjCDeclQualifier_Byref

CXObjCDeclQualifier_In

CXObjCDeclQualifier_Inout

CXObjCDeclQualifier_None

CXObjCDeclQualifier_Oneway

CXObjCDeclQualifier_Out

CXObjCPropertyAttr_assign

CXObjCPropertyAttr_atomic

CXObjCPropertyAttr_copy

CXObjCPropertyAttr_getter

CXObjCPropertyAttr_noattr

CXObjCPropertyAttr_nonatomic

CXObjCPropertyAttr_readonly

CXObjCPropertyAttr_readwrite

CXObjCPropertyAttr_retain

CXObjCPropertyAttr_setter

CXObjCPropertyAttr_strong

CXObjCPropertyAttr_unsafe_unretained

CXObjCPropertyAttr_weak

CXReparse_None

CXSaveTranslationUnit_None

CXTranslationUnit_CXXChainedPCH

CXTranslationUnit_CacheCompletionResults

CXTranslationUnit_CreatePreambleOnFirstParse

CXTranslationUnit_DetailedPreprocessingRecord

CXTranslationUnit_ForSerialization

CXTranslationUnit_IncludeBriefCommentsInCodeCompletion

CXTranslationUnit_Incomplete

CXTranslationUnit_None

CXTranslationUnit_PrecompiledPreamble

CXTranslationUnit_SkipFunctionBodies

Functions

clang_BlockCommandComment_getArgText

clang_BlockCommandComment_getCommandName

clang_BlockCommandComment_getNumArgs

clang_BlockCommandComment_getParagraph

clang_CXCursorSet_contains

clang_CXCursorSet_insert

clang_CXIndex_getGlobalOptions

clang_CXIndex_setGlobalOptions

clang_CXXField_isMutable

clang_CXXMethod_isConst

clang_CXXMethod_isPureVirtual

clang_CXXMethod_isStatic

clang_CXXMethod_isVirtual

clang_Comment_getChild

clang_Comment_getKind

clang_Comment_getNumChildren

clang_Comment_isWhitespace

clang_CompilationDatabase_dispose

clang_CompilationDatabase_fromDirectory

clang_CompilationDatabase_getAllCompileCommands

clang_CompilationDatabase_getCompileCommands

clang_CompileCommand_getArg

clang_CompileCommand_getDirectory

clang_CompileCommand_getFilename

clang_CompileCommand_getMappedSourceContent

clang_CompileCommand_getMappedSourcePath

clang_CompileCommand_getNumArgs

clang_CompileCommands_dispose

clang_CompileCommands_getCommand

clang_CompileCommands_getSize

clang_Cursor_getArgument

clang_Cursor_getBriefCommentText

clang_Cursor_getCXXManglings

clang_Cursor_getCommentRange

clang_Cursor_getMangling

clang_Cursor_getModule

clang_Cursor_getNumArguments

clang_Cursor_getNumTemplateArguments

clang_Cursor_getObjCDeclQualifiers

clang_Cursor_getObjCPropertyAttributes

clang_Cursor_getObjCSelectorIndex

clang_Cursor_getOffsetOfField

clang_Cursor_getParsedComment

clang_Cursor_getRawCommentText

clang_Cursor_getReceiverType

clang_Cursor_getSpellingNameRange

clang_Cursor_getStorageClass

clang_Cursor_getTemplateArgumentKind

clang_Cursor_getTemplateArgumentType

clang_Cursor_getTemplateArgumentUnsignedValue

clang_Cursor_getTemplateArgumentValue

clang_Cursor_getTranslationUnit

clang_Cursor_isAnonymous

clang_Cursor_isBitField

clang_Cursor_isDynamicCall

clang_Cursor_isNull

clang_Cursor_isObjCOptional

clang_Cursor_isVariadic

clang_File_isEqual

clang_FullComment_getAsHTML

clang_FullComment_getAsXML

clang_HTMLStartTagComment_isSelfClosing

clang_HTMLStartTag_getAttrName

clang_HTMLStartTag_getAttrValue

clang_HTMLStartTag_getNumAttrs

clang_HTMLTagComment_getAsString

clang_HTMLTagComment_getTagName

clang_IndexAction_create

clang_IndexAction_dispose

clang_InlineCommandComment_getArgText

clang_InlineCommandComment_getCommandName

clang_InlineCommandComment_getNumArgs

clang_InlineCommandComment_getRenderKind

clang_InlineContentComment_hasTrailingNewline

clang_Location_isFromMainFile

clang_Location_isInSystemHeader

clang_Module_getASTFile

clang_Module_getFullName

clang_Module_getName

clang_Module_getNumTopLevelHeaders

clang_Module_getParent

clang_Module_getTopLevelHeader

clang_Module_isSystem

clang_ParamCommandComment_getDirection

clang_ParamCommandComment_getParamIndex

clang_ParamCommandComment_getParamName

clang_ParamCommandComment_isDirectionExplicit

clang_ParamCommandComment_isParamIndexValid

clang_Range_isNull

clang_TParamCommandComment_getDepth

clang_TParamCommandComment_getIndex

clang_TParamCommandComment_getParamName

clang_TParamCommandComment_isParamPositionValid

clang_TextComment_getText

clang_Type_getAlignOf

clang_Type_getCXXRefQualifier

clang_Type_getClassType

clang_Type_getNumTemplateArguments

clang_Type_getOffsetOf

clang_Type_getSizeOf

clang_Type_getTemplateArgumentAsType

clang_Type_visitFields

clang_VerbatimBlockLineComment_getText

clang_VerbatimLineComment_getText

clang_annotateTokens

clang_codeCompleteAt

clang_codeCompleteGetContainerKind

clang_codeCompleteGetContainerUSR

clang_codeCompleteGetContexts

clang_codeCompleteGetDiagnostic

clang_codeCompleteGetNumDiagnostics

clang_codeCompleteGetObjCSelector

clang_constructUSR_ObjCCategory

clang_constructUSR_ObjCClass

clang_constructUSR_ObjCIvar

clang_constructUSR_ObjCMethod

clang_constructUSR_ObjCProperty

clang_constructUSR_ObjCProtocol

clang_createCXCursorSet

clang_createIndex

clang_createTranslationUnit

clang_createTranslationUnit2

clang_createTranslationUnitFromSourceFile

clang_defaultCodeCompleteOptions

clang_defaultDiagnosticDisplayOptions

clang_defaultEditingTranslationUnitOptions

clang_defaultReparseOptions

clang_defaultSaveOptions

clang_disposeCXCursorSet

clang_disposeCXPlatformAvailability

clang_disposeCXTUResourceUsage

clang_disposeCodeCompleteResults

clang_disposeDiagnostic

clang_disposeDiagnosticSet

clang_disposeIndex

clang_disposeOverriddenCursors

clang_disposeSourceRangeList

clang_disposeString

clang_disposeStringSet

clang_disposeTokens

clang_disposeTranslationUnit

clang_enableStackTraces

clang_equalCursors

clang_equalLocations

clang_equalRanges

clang_equalTypes

clang_executeOnThread

clang_findIncludesInFile

clang_findReferencesInFile

clang_formatDiagnostic

clang_free

clang_getArgType

clang_getArrayElementType

clang_getArraySize

clang_getCString

clang_getCXTUResourceUsage

clang_getCXXAccessSpecifier

clang_getCanonicalCursor

clang_getCanonicalType

clang_getChildDiagnostics

clang_getClangVersion

clang_getCompletionAnnotation

clang_getCompletionAvailability

clang_getCompletionBriefComment

clang_getCompletionChunkCompletionString

clang_getCompletionChunkKind

clang_getCompletionChunkText

clang_getCompletionNumAnnotations

clang_getCompletionParent

clang_getCompletionPriority

clang_getCursor

clang_getCursorAvailability

clang_getCursorCompletionString

clang_getCursorDefinition

clang_getCursorDisplayName

clang_getCursorExtent

clang_getCursorKind

clang_getCursorKindSpelling

clang_getCursorLanguage

clang_getCursorLexicalParent

clang_getCursorLinkage

clang_getCursorLocation

clang_getCursorPlatformAvailability

clang_getCursorReferenceNameRange

clang_getCursorReferenced

clang_getCursorResultType

clang_getCursorSemanticParent

clang_getCursorSpelling

clang_getCursorType

clang_getCursorUSR

clang_getCursorVisibility

clang_getDeclObjCTypeEncoding

clang_getDefinitionSpellingAndExtent

clang_getDiagnostic

clang_getDiagnosticCategory

clang_getDiagnosticCategoryName

clang_getDiagnosticCategoryText

clang_getDiagnosticFixIt

clang_getDiagnosticInSet

clang_getDiagnosticLocation

clang_getDiagnosticNumFixIts

clang_getDiagnosticNumRanges

clang_getDiagnosticOption

clang_getDiagnosticRange

clang_getDiagnosticSetFromTU

clang_getDiagnosticSeverity

clang_getDiagnosticSpelling

clang_getElementType

clang_getEnumConstantDeclUnsignedValue

clang_getEnumConstantDeclValue

clang_getEnumDeclIntegerType

clang_getExpansionLocation

clang_getFieldDeclBitWidth

clang_getFile

clang_getFileLocation

clang_getFileName

clang_getFileTime

clang_getFileUniqueID

clang_getFunctionTypeCallingConv

clang_getIBOutletCollectionType

clang_getIncludedFile

clang_getInclusions

clang_getInstantiationLocation

clang_getLocation

clang_getLocationForOffset

clang_getModuleForFile

clang_getNullCursor

clang_getNullLocation

clang_getNullRange

clang_getNumArgTypes

clang_getNumCompletionChunks

clang_getNumDiagnostics

clang_getNumDiagnosticsInSet

clang_getNumElements

clang_getNumOverloadedDecls

clang_getOverloadedDecl

clang_getOverriddenCursors

clang_getPointeeType

clang_getPresumedLocation

clang_getRange

clang_getRangeEnd

clang_getRangeStart

clang_getRemappings

clang_getRemappingsFromFileList

clang_getResultType

clang_getSkippedRanges

clang_getSpecializedCursorTemplate

clang_getSpellingLocation

clang_getTUResourceUsageName

clang_getTemplateCursorKind

clang_getTokenExtent

clang_getTokenKind

clang_getTokenLocation

clang_getTokenSpelling

clang_getTranslationUnitCursor

clang_getTranslationUnitSpelling

clang_getTypeDeclaration

clang_getTypeKindSpelling

clang_getTypeSpelling

clang_getTypedefDeclUnderlyingType

clang_hashCursor

clang_indexLoc_getCXSourceLocation

clang_indexLoc_getFileLocation

clang_indexSourceFile

clang_indexSourceFileFullArgv

clang_indexTranslationUnit

clang_index_getCXXClassDeclInfo

clang_index_getClientContainer

clang_index_getClientEntity

clang_index_getIBOutletCollectionAttrInfo

clang_index_getObjCCategoryDeclInfo

clang_index_getObjCContainerDeclInfo

clang_index_getObjCInterfaceDeclInfo

clang_index_getObjCPropertyDeclInfo

clang_index_getObjCProtocolRefListInfo

clang_index_isEntityObjCContainerKind

clang_index_setClientContainer

clang_index_setClientEntity

clang_isAttribute

clang_isConstQualifiedType

clang_isCursorDefinition

clang_isDeclaration

clang_isExpression

clang_isFileMultipleIncludeGuarded

clang_isFunctionTypeVariadic

clang_isInvalid

clang_isPODType

clang_isPreprocessing

clang_isReference

clang_isRestrictQualifiedType

clang_isStatement

clang_isTranslationUnit

clang_isUnexposed

clang_isVirtualBase

clang_isVolatileQualifiedType

clang_loadDiagnostics

clang_parseTranslationUnit

clang_parseTranslationUnit2

clang_parseTranslationUnit2FullArgv

clang_remap_dispose

clang_remap_getFilenames

clang_remap_getNumFiles

clang_reparseTranslationUnit

clang_saveTranslationUnit

clang_sortCodeCompletionResults

clang_toggleCrashRecovery

clang_tokenize

clang_visitChildren

load
Loads a libclang shared library for use in the current thread.

unload
Unloads the libclang shared library in use in the current thread.

Type Definitions

CXClientData

CXCursorVisitor

CXFieldVisitor

CXInclusionVisitor